# internal/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: MajorGTM API
  version: "1.57.1"
  description: API for MajorGTM recorded updates and async video messaging.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://rec.majorgtm.com
    description: Production
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthResponse:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
paths:
...
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Uses the refresh_token cookie to issue a new access token. Rate limited to 0.5 requests per second with a burst of 5.
      operationId: refreshToken
      responses:
        "200":
          description: Token refreshed
          headers:
            Set-Cookie:
              description: refresh_token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or missing refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...